+++ openfisca_france/model/prestations/prestations_locales/paris/paris.xml
@@ -84,7 +84,7 @@
     </CODE>
   </NODE>
   <NODE code="plfm" description="Paris logement familles monoparentales">
-     <CODE code="premier_plafond_plfm" description="1er Plafond des ressources pour l'aide paris forfait famille " format="float" type="monetary">
+    <CODE code="premier_plafond_plfm" description="1er Plafond des ressources pour l'aide paris forfait famille " format="float" type="monetary">
       <VALUE deb="2015-01-01" fin="2016-01-01" valeur="1140"/>
     </CODE>
     <CODE code="deuxieme_plafond_plfm" description="2ème Plafond des ressources pour l'aide paris forfait famille" format="float" type="monetary">
@@ -129,6 +129,32 @@
     <CODE code="montant_aide_cs" description="Montant de l'aide Complémentaire Santé " format="float" type="monetary">
       <VALUE deb="2015-01-01" fin="2016-01-31" valeur="468"/>
       <VALUE deb="2016-02-01" fin="2016-12-31" valeur="468"/>
+    </CODE>
+  </NODE>
+  <NODE code="paris_logement_familles" description="Paris Logement Famille">
+  <CODE code="plafond_haut_3enf" description="Plafond haut de PLF pour les familles à 3 enfants, aussi plafond de PLF avec enfant handicapé" format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="5000" />
+    </CODE>
+    <CODE code="plafond_bas_3enf" description="Plafond bas de PLF pour les familles à 3 enfants" format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="3000" />
+    </CODE>
+    <CODE code="plafond_2enf" description="Plafond de PLF pour les familles à deux enfants." format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="2000" />
+    </CODE>
+    <CODE code="montant_haut_3enf" description="Montant haut PLF pour famille à 3 enfants." format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="128" />
+    </CODE>
+    <CODE code="montant_bas_3enf" description="Montant bas PLF pour famille à 3 enfants." format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="84" />
+    </CODE>
+    <CODE code="montant_2enf" description="Montant PLF pour les familles à deux enfants." format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="116" />
+    </CODE>
+    <CODE code="montant_haut_enf_sup" description="Montant haut sup par enfant" format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="41" />
+    </CODE>
+    <CODE code="montant_bas_enf_sup" description="Montant bas sup par enfant." format="float" type="monetary">
+      <VALUE deb="2014-01-01" fin="2015-12-31" valeur="21" />
     </CODE>
   </NODE>
 </NODE>
